meta-hp: linux-hp-tenderloin: switch to a prebuilt kernel
We have to use a prebuilt kernel as building the kernel for tenderloin with gcc 4.8 isn't
possible right now. When starting a kernel build with gcc 4.8 it crashes early at boot
time and we don't have a fix for this yet. For now we build the kernel as part of android
and combine it with the correct initramfs here.

Signed-off-by: Simon Busch <morphis@gravedo.de>